Replaces favorite table view with with collection view
