--- itesm_event_listing/FavoriteTableViewCell.swift
@@ -1,28 +1,0 @@
-//
-//  FavoriteTableViewCell.swift
-//  itesm_event_listing
-//
-//  Created by Andrés Villanueva on 4/9/18.
-//  Copyright © 2018 Rene Garcia Saenz. All rights reserved.
-//
-
-import UIKit
-
-class FavoriteTableViewCell: UITableViewCell {
-
-    @IBOutlet weak var eventName: UILabel!
-    @IBOutlet weak var eventLocation: UILabel!
-    @IBOutlet weak var eventImage: UIImageView!
-    
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        // Initialization code
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-        // Configure the view for the selected state
-    }
-
-}
